LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/fill-and-stroke-styles/2d.fillStyle.parse.css-color-4-hsl-5-expected.txt

2d.fillStyle.parse.css-color-4-hsl-5


FAIL OffscreenCanvas test: 2d.fillStyle.parse.css-color-4-hsl-5 assert_equals: Red channel of the pixel at (50, 25) expected 0 but got 255
